Backend/myshop/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-allauth==65.5.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
